Move keywords to crypto-validation
